requirements.txt
tensorflow-cpu>=2.12
numpy>=1.21,<1.24
Pillow>=9.0,<10.0
matplotlib>=3.5,<3.8
flask>=2.0,<3.0
flask-cors>=3.0,<4.0
werkzeug>=2.0,<3.0
